refactor(file-memory-store): split consolidation into planning/validation/execution modules
Move consolidation out of the monolithic file-memory-store.ts (~1300 lines)
into a dedicated consolidation/ directory, separating the three concerns
along their natural seams:

- consolidation/plan.ts     — the plan schema, types, and plan extraction
- consolidation/planner.ts  — all model interaction (prompt + revise-retry)
- consolidation/validate.ts — pure guardrail functions over (plan, files)
- consolidation/execute.ts  — deterministic execution against storage

Shared primitives that both the store and the consolidation modules depend
on (frontmatter format, reserved changelog path, path-identity rule,
concurrency-bounded map) move to a sibling internal.ts, so consolidation/
never imports back into the store and no import cycle forms.

file-memory-store.ts retains the public consolidate() entry point and the
run orchestration, delegating planning, validation, and execution to the
new modules. Behavior-preserving: no logic, guardrail, prompt, or error
message changed; the executor functions take storage as a parameter rather
than reading it off the instance. Test files are unchanged and the full
suite passes, verifying the reorganization end to end.
